refactor : BookDetailActivity 이동
