+++ Android/app/src/main/java/com/booktree/ui/book/bookDetail/BookDetailViewModel.java
@@ -1,4 +1,4 @@
-package com.booktree;
+package com.booktree.ui.book.bookDetail;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
