+++ Android/app/src/main/java/com/booktree/ui/book/bookList/BookListAdapter.java
@@ -1,23 +1,16 @@
 package com.booktree.ui.book.bookList;
 
 import android.content.Context;
-import android.content.Intent;
 import android.view.LayoutInflater;
 import android.view.View;
-import android.view.View.OnClickListener;
 import android.view.ViewGroup;
-import android.widget.ImageView;
-import android.widget.TextView;
 import androidx.annotation.NonNull;
-import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.recyclerview.widget.RecyclerView;
 import com.booktree.model.Documents;
-import com.booktree.BookDetailActivity;
 import com.booktree.R;
 import com.booktree.ui.book.bookList.Viewholder.BasicViewHolder;
 import com.booktree.ui.book.bookList.Viewholder.ToBookInfoViewHolder;
 import com.booktree.ui.book.bookList.Viewholder.getBookInfoViewholder;
-import com.bumptech.glide.Glide;
 import java.util.ArrayList;
 import java.util.Arrays;
 import org.jetbrains.annotations.NotNull;
