+++ Android/app/src/main/java/com/booktree/ui/book/bookList/Viewholder/BasicViewHolder.java
@@ -1,14 +1,11 @@
 package com.booktree.ui.book.bookList.Viewholder;
 
 import android.content.Context;
-import android.content.Intent;
-import android.view.LayoutInflater;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
 import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.recyclerview.widget.RecyclerView;
-import com.booktree.BookDetailActivity;
 import com.booktree.R;
 import com.booktree.model.Documents;
 import com.bumptech.glide.Glide;
